Refactor height measurement
