+++ src/Measurements.Core/Height/Bones.cs
@@ -1,0 +1,10 @@
+namespace Measurements.Height
+{
+    internal enum Bones
+    {
+        /// <summary>
+        /// Top of the head
+        /// </summary>
+        N_Head_top,
+    }
+}
