+++ src/Measurements.Core/Height/Gui.cs
@@ -1,8 +1,9 @@
-﻿using KKAPI.Maker;
+using KKAPI.Maker;
+using Measurements.Gui;
 
-namespace Measurements.Gui
+namespace Measurements.Height
 {
-    internal class HeightGui : TextGui
+    internal class Gui : TextGui
     {
         public override void Initialize(
             MakerCategory category,
@@ -10,7 +11,7 @@ namespace Measurements.Gui
             RegisterSubCategoriesEvent e
         ) => InitializeInternal("Height", category, plugin, e);
 
-        public override void Update(MeasurementsData data, MeasurementsController controller)
+        protected override void UpdateInternal(MeasurementsData data, MeasurementsController controller)
         {
             if (controller.UseMetricUnits) SetText($"{data.Height:N1} cm");
             else
